refactor(servers): move log components to data directory for better organization
The log-related components and pages were moved from the 'logs' directory to 'data' to better reflect their purpose and improve project structure. This includes:
- Moving login log and build/destroy log components
- Updating the navigation menu to reflect the new path
- Maintaining all existing functionality while reorganizing files
