+++ app/[locale]/(main)/servers/[id]/data/build-destroy-list-footer.tsx
