+++ app/[locale]/(main)/servers/[id]/data/server-build-log-card.tsx
